# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LeaderboardEntry {
  id: number;
  playerName: string;
  trackId: string;
  totalTimeMs: number;
  bestLapMs: number;
  lapsCompleted: number;
  createdAt: string;
}

export interface ScoreInput {
  /**
     * @minLength 1
     * @maxLength 20
     */
  playerName: string;
  trackId: string;
  totalTimeMs: number;
  bestLapMs: number;
  lapsCompleted: number;
}

export interface AuthInput {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthResponse {
  token: string;
  userId: number;
  username: string;
}

export type PlayerProfilePersonalBests = {[key: string]: number};

export interface PlayerProfile {
  userId: number;
  username: string;
  level: number;
  xp: number;
  totalRaces: number;
  personalBests: PlayerProfilePersonalBests;
}

export interface ProfileUpdate {
  xpGained: number;
  trackId: string;
  bestLapMs: number;
  totalTimeMs: number;
  lapsCompleted: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetLeaderboardParams = {
limit?: number;
/**
 * @nullable
 */
trackId?: string | null;
};

export type GetPersonalBestParams = {
playerName: string;
trackId: string;
};