# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: leaderboard
  - name: auth
  - name: profile
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get top leaderboard scores
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: trackId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
    post:
      operationId: submitScore
      tags: [leaderboard]
      summary: Submit a race score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreInput"
      responses:
        "201":
          description: Score submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardEntry"
  /leaderboard/personal-best:
    get:
      operationId: getPersonalBest
      tags: [leaderboard]
      summary: Get personal best for a player
      parameters:
        - name: playerName
          in: query
          required: true
          schema:
            type: string
        - name: trackId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Personal best entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardEntry"
        "404":
          description: No entry found
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Validate token and get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get player profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile after a race
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    LeaderboardEntry:
      type: object
      properties:
        id:
          type: integer
        playerName:
          type: string
        trackId:
          type: string
        totalTimeMs:
          type: integer
        bestLapMs:
          type: integer
        lapsCompleted:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, playerName, trackId, totalTimeMs, bestLapMs, lapsCompleted, createdAt]
    ScoreInput:
      type: object
      properties:
        playerName:
          type: string
          minLength: 1
          maxLength: 20
        trackId:
          type: string
        totalTimeMs:
          type: integer
        bestLapMs:
          type: integer
        lapsCompleted:
          type: integer
      required: [playerName, trackId, totalTimeMs, bestLapMs, lapsCompleted]
    AuthInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        password:
          type: string
          minLength: 6
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        userId:
          type: integer
        username:
          type: string
      required: [token, userId, username]
    PlayerProfile:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        level:
          type: integer
        xp:
          type: integer
        totalRaces:
          type: integer
        personalBests:
          type: object
          additionalProperties:
            type: integer
      required: [userId, username, level, xp, totalRaces, personalBests]
    ProfileUpdate:
      type: object
      properties:
        xpGained:
          type: integer
        trackId:
          type: string
        bestLapMs:
          type: integer
        totalTimeMs:
          type: integer
        lapsCompleted:
          type: integer
      required: [xpGained, trackId, bestLapMs, totalTimeMs, lapsCompleted]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]