# lib/api-zod/src/generated/types/authInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthInput {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
  /** @minLength 6 */
  password: string;
}