# lib/api-zod/src/generated/types/playerProfile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerProfilePersonalBests } from './playerProfilePersonalBests';

export interface PlayerProfile {
  userId: number;
  username: string;
  level: number;
  xp: number;
  totalRaces: number;
  personalBests: PlayerProfilePersonalBests;
}