# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  xpGained: number;
  trackId: string;
  bestLapMs: number;
  totalTimeMs: number;
  lapsCompleted: number;
}